fix: Swap sendBeacon for fetch (keepalive=true) in ServerStream unload
Change that suffices as a satisfactory replacement for sendBeacon which is restricting
the method and content type of log request on page unload.

See https://stackoverflow.com/questions/55421459/http-request-beforeunload-sendbeacon-vs-img-src

Note that it might seem odd that regualar requests are sent with XMLHttpRequest but the unload is set
with Fetch. This is so maximum browser compatibility for 95% operation is preserved. For non-blocking
unload requests, Fetch seems to be the overall best option.