fix: Improve type defnitions to match the actual js
I'd like to propose a few changes to the type definitions of the main package. These would allow the package to better integrate in TypeScript projects.

type LoggerOptions = { .... [key: string]: any; }

Allows any additional properties to be added to the logged object as documented here in node-bunyan. As this also works for browser-buynan I added the respective documentation from node-bunyan.

The definition for the first argument for Logger#child is wrong, as it doesn't expect LoggerOptions. Actually it expects LoggerOptions without the name property and an optional childName property and again any other additional properties. Here I added the type ChildLoggerOptions and referenced it as the type for the first argument of Logger#child

Lastly I turned the interface for Logger into an exported class declaration, as it's actually a class that's exported from the module.

These should not be a breaking changes.